test(chat): cover math copy in Gecko DOM
